BREAKING CHANGE: deprecated front matter class
