+++ src/Content/Markdown.php
@@ -20,11 +20,6 @@ class Markdown
 {
     private string $fileContent = '';
 
-    /**
-     * @var FrontMatter
-     */
-    private FrontMatter $frontMatter;
-
     private string $body = '';
 
     public static function for(File $file, FileSystemInterface $in): Markdown
@@ -90,7 +85,7 @@ class Markdown
 
                 } else {
                     $b = $template::create(
-                        $this->frontMatter(),
+                        $this->file(),
                         $this->fileSystem()
                     );
 
@@ -106,15 +101,15 @@ class Markdown
     /**
      * @todo: test
      */
-    public function frontMatter(): FrontMatter
-    {
-        if (! isset($this->frontMatter)) {
-            $frontMatter = self::markdownConverter()
-                ->getFrontMatter($this->fileContent());
-            $this->frontMatter = FrontMatter::init($frontMatter);
-        }
-        return $this->frontMatter;
-    }
+    // public function frontMatter(): FrontMatter
+    // {
+    //     if (! isset($this->frontMatter)) {
+    //         $frontMatter = self::markdownConverter()
+    //             ->getFrontMatter($this->fileContent());
+    //         $this->frontMatter = FrontMatter::init($frontMatter);
+    //     }
+    //     return $this->frontMatter;
+    // }
 
     public function body(): string
     {
@@ -143,53 +138,53 @@ class Markdown
 //         return implode(' | ', $titles);
 //     }
 
-    public function description(): string
-    {
-        if ($this->frontMatter()->hasMember('description')) {
-            $description = $this->frontMatter()->description();
-
-        } else {
-            $body = $this->body();
-            $description = preg_filter(
-                ["/#(.*)\n/", "/{!!(.*)!!}/"],
-                ['', ''],
-                $body
-            );
-            if (is_string($description)) {
-                $parts = explode("\n", $description);
-                $parts = array_filter($parts);
-                $description = implode(' ', $parts);
-
-            } else {
-                // TODO: Doesn't guarantee meta description content.
-                //       Log??
-                $description = $body;
-
-            }
-        }
-
-        $description = htmlentities(substr($description, 0, 200));
-
-        $parts = explode('. ', $description);
-        $description = '';
-        foreach ($parts as $part) {
-            $d = $part;
-            if (strlen($description) > 0) {
-                $d = $description . '. ' . $part;
-            }
-
-            $proposedLength = strlen($d);
-            if ($proposedLength >= 200) {
-                $ps = explode('. ', $d);
-                array_pop($ps);
-                $description = implode('. ', $ps) . '.';
-                break;
-            }
-            $description = $d;
-        }
-
-        return $description;
-    }
+//     public function description(): string
+//     {
+//         if ($this->frontMatter()->hasMember('description')) {
+//             $description = $this->frontMatter()->description();
+//
+//         } else {
+//             $body = $this->body();
+//             $description = preg_filter(
+//                 ["/#(.*)\n/", "/{!!(.*)!!}/"],
+//                 ['', ''],
+//                 $body
+//             );
+//             if (is_string($description)) {
+//                 $parts = explode("\n", $description);
+//                 $parts = array_filter($parts);
+//                 $description = implode(' ', $parts);
+//
+//             } else {
+//                 // TODO: Doesn't guarantee meta description content.
+//                 //       Log??
+//                 $description = $body;
+//
+//             }
+//         }
+//
+//         $description = htmlentities(substr($description, 0, 200));
+//
+//         $parts = explode('. ', $description);
+//         $description = '';
+//         foreach ($parts as $part) {
+//             $d = $part;
+//             if (strlen($description) > 0) {
+//                 $d = $description . '. ' . $part;
+//             }
+//
+//             $proposedLength = strlen($d);
+//             if ($proposedLength >= 200) {
+//                 $ps = explode('. ', $d);
+//                 array_pop($ps);
+//                 $description = implode('. ', $ps) . '.';
+//                 break;
+//             }
+//             $description = $d;
+//         }
+//
+//         return $description;
+//     }
 
     public function file(): File
     {
